Redo DLS test_snapshot_normal integration test
This test was previously using DLS-specific PVs. It has been switched
over to using the soft IOC to fix this.

For clarity, it has been split in two. The first checks whether a very
basic request file can be processed properly, while the second checks
the behaviour of modifiers and comments.