Remove reference to DLS Control Room from tests
References to the DLS Control Room were included as we needed to ensure
that pyburt was capable of handling all PV types used there.

I believe this information is better kept outside of the module, as it
is not necessary to understand how the tests work and is unclear for
third parties who wish to contribute to pyburt.